Add connections and a trial data model.
Add data-write-permission flags to the message-editing view and
window-controller.  Add bindings between the two during window setup.
This requires making the window-controller the delegate of its window.

In the message document class, factor out the code to create
message-editing window to a separate function.  Add a temporary data
model for messages (to test the view classes later).

Add an in-program symbol for the message document class’s UTI.