Add CoreData-based message framework.
Add framework with a Core Data model for the Internet Message Format,
with (altered) generated managed-object subclasses.  The framework
includes a protocol for IMF header fields, support types, an exclusive
addition to String, and test cases.

Replace the temporary message and header-field classes with the ones
from the new framework. Add loading of a Core Data stack to the
document subclass, adding copies of the super class’s initializers to
propagate errors.. Change the connection of the message data between
the document subclass and window controller to Cocoa Bindings.

Workaround the bug of NSArrayController not being able to use a
NSOrderedSet by providing a ValueTransformer class that trades from a
NSOrderedSet to a NSArray.