Replace main document window.
Remove the old window’s storyboard, window controller class, and view
controller class.  Add new storyboard and view controller class (using
default window controller class).  The new window shows either the
table view for the header or the text view for the body, but not both
at the same time, using a tab view to switch.  Add custom buttons to
switch between the tabs.  Add buttons to add or remove header fields to
the table view.

Add commands to the document class to add or remove a message body.
Move the replacement window creation method to an extension.  Move the
convenience initializers to an extension.

Create a pseudo-property on the message data class that exposes the
header collection property as an array (instead of just an
ordered-set).  Remove the now-supplanted value-transformer class.