Add file-loading to document class.
Add iterator type for traversing a block of data, returning lines (as
strings).  Add tests for it and its implementation types.

Add method to parse a text line to its role in an e-mail; as a
type-level method of the enumeration type of the parsing results.  Add
tests.

Add a (possibly-threaded) task type to convert a block of data into a
parsed text of an e-mail.  Add tests.

Change top-level saving method in the document class to ensure
error-completion is called in the main thread.  Change the
asynchronous-writing indicator and bottom-level saving methods’ way of
checking the UTI.  Change how the message is copied in the background
for the bottom-level saving method.

Add an implementation of loading at the document class using the new
message-parsing code.  Enable concurrent loading.  Remove the default
data from new (not loaded) documents.