Replace main document window, again.
Remove the document window's view controller class and storyboard.  Add
new storyboard (reusing the file name) and window controller class as
replacements.  The new window shows the header and body at the same
time, stacked in a 2-way split view.  The header data is still
displayed in a table view.  The body's state is shown in one of two
tabs: either the missing state with a button with an action to add a
blank body, or the non-null state with a text view to edit the body
string.  Move the buttons for adding or removing a header field to the
toolbar.