Add menu commands to append/strip message bodies.
Move the actions for adding or removing a body string from a message
from the document subclass to its window controller subclass.  Change
their implementation to pass through the window controller’s message
object controller.

Add a menu and menu items for the two actions.  Add an override to the
window controller to determine when each action is active.