Add font preferences, and use them.
Add file to store keys of app defaults.  This first set of defaults are
the font settings for lists and text blocks, both on-screen and for
printing.  Include initial values to use as part of the registration
domain.

Include conversion functions between the font-specifier type and a
defaults-compatible type.  Extend the system defaults class with
properties for the new app defaults.

Expand the application delegate to load the new defaults on startup.

Use the on-screen text-block font for the document window’s text view.
Widen the window and its components so at least 80 characters are
visible per line by default.  Use all four font settings to part of
attributed-string conversion in the document class during printing and
export-as-RTF.