Tweak open/save priorities.
Refine in the Info.plist file the handling rank of opening UTF-8 and
ASCII e-mail files.  Specify that these file types are not packages.

Although reading a particular UTI type should include its refined
types, writing should include the UTI’s generalized types instead.
Switch the direction of UTI checking during the document subclass’s
writing to match this creed.  Note that a consequence of this reversal
is that which UTI to use for writing has to be checked from most
expressive on down.