Add menu items to add & remove header fields.
Add items under the “Message” menu for the actions to add and remove
header fields.  Refine the validation of the remove-header-field action
to display the count of fields affected in its menu item’s title (if
any).  Add a default validation of this action in the app delegate to
restore the original menu item title.  Add localization files to enable
the new title string values.

Update the message window’s header display to allow multiple selection.